changing the model of workout and get the app startable

+++ app/src/main/java/com/example/gymlog/domain/model/workout/WorkoutResult.kt
@@ -8,6 +8,6 @@ data class WorkoutResult(
     val workTime: Int?,
     val position: Int = 0,
     val sequenceInGymDay: Int,
-    val timeFromStart: Int,
+    val timeFromStart: Long,
     val workoutDateTime: String // Формат "yyyy.MM.dd HH:mm"
 )
